requirements.txt
pandas
ta
ccxt
python-telegram-bot
schedule